Switch ID type int to Long for consistency

+++ src/main/java/com/jgcosmetics/store/repository/UserRepo.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
 
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
     User findByUsername(String username);
     User findByEmail(String email);
 }
